increase some dependencies in Rust including FRB and generate code

+++ packages/wasm_run/native/src/config.rs
@@ -179,7 +179,7 @@ impl From<ModuleConfig> for wasmtime::Config {
             wtc.wasm_backtrace.map(|v| config.wasm_backtrace(v));
             wtc.native_unwind_info.map(|v| config.native_unwind_info(v));
             // wtc.epoch_interruption.map(|v| config.epoch_interruption(v));
-            wtc.max_wasm_stack.map(|v| config.max_wasm_stack(v));
+            wtc.max_wasm_stack.map(|v| config.max_wasm_stack(v as usize));
             wtc.wasm_simd.map(|v| config.wasm_simd(v));
             wtc.wasm_relaxed_simd.map(|v| config.wasm_relaxed_simd(v));
             wtc.relaxed_simd_deterministic
@@ -235,7 +235,7 @@ pub struct ModuleConfigWasmi {
     /// The limits set on the value stack and call stack.
     pub stack_limits: Option<WasiStackLimits>,
     /// The amount of Wasm stacks to keep in cache at most.
-    pub cached_stacks: Option<usize>,
+    pub cached_stacks: Option<i64>,
     /// Is `true` if the `mutable-global` Wasm proposal is enabled.
     pub mutable_global: Option<bool>,
     /// Is `true` if the `sign-extension` Wasm proposal is enabled.
@@ -258,11 +258,11 @@ pub struct ModuleConfigWasmi {
 #[derive(Debug, Copy, Clone)]
 pub struct WasiStackLimits {
     /// The initial value stack height that the Wasm stack prepares.
-    pub initial_value_stack_height: usize,
+    pub initial_value_stack_height: i64,
     /// The maximum value stack height in use that the Wasm stack allows.
-    pub maximum_value_stack_height: usize,
+    pub maximum_value_stack_height: i64,
     /// The maximum number of nested calls that the Wasm stack allows.
-    pub maximum_recursion_depth: usize,
+    pub maximum_recursion_depth: i64,
 }
 
 #[cfg(not(feature = "wasmtime"))]
@@ -293,7 +293,7 @@ pub struct ModuleConfigWasmtime {
     // TODO: pub wasm_backtrace_details: WasmBacktraceDetails, // Or WASMTIME_BACKTRACE_DETAILS env var
     //
     // TODO: pub epoch_interruption: Option<bool>, // vs consume_fuel
-    pub max_wasm_stack: Option<usize>,
+    pub max_wasm_stack: Option<i64>,
     /// Whether or not to enable the `threads` WebAssembly feature.
     /// This includes atomics and shared memory as well.
     /// This is not enabled by default.
